HAAR-1930: refactor POST /clients/client-credentials/add to /base-clients
