+++ src/main/kotlin/uk/gov/justice/digital/hmpps/authorizationserver/resource/ClientsController.kt
@@ -1,6 +1,7 @@
 package uk.gov.justice.digital.hmpps.authorizationserver.resource
 
 import com.microsoft.applicationinsights.TelemetryClient
+import jakarta.validation.Valid
 import org.springframework.core.convert.ConversionService
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -10,6 +11,7 @@ import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.ResponseStatus
 import uk.gov.justice.digital.hmpps.authorizationserver.config.AuthenticationFacade
@@ -18,6 +20,7 @@ import uk.gov.justice.digital.hmpps.authorizationserver.data.model.ClientType
 import uk.gov.justice.digital.hmpps.authorizationserver.service.ClientDetail
 import uk.gov.justice.digital.hmpps.authorizationserver.service.ClientFilter
 import uk.gov.justice.digital.hmpps.authorizationserver.service.ClientsService
+import uk.gov.justice.digital.hmpps.authorizationserver.service.SaveClientService
 import uk.gov.justice.digital.hmpps.authorizationserver.service.SortBy
 import java.time.Instant
 
@@ -27,6 +30,7 @@ class ClientsController(
   private val conversionService: ConversionService,
   private val telemetryClient: TelemetryClient,
   private val authenticationFacade: AuthenticationFacade,
+  private val saveClientService: SaveClientService,
 ) {
 
   @GetMapping("base-clients")
@@ -38,6 +42,19 @@ class ClientsController(
     @RequestParam clientType: ClientType? = null,
   ): ResponseEntity<Any> {
     return ResponseEntity.ok(AllClientsResponse(clientsService.retrieveAllClients(sort, ClientFilter(grantType = grantType, role = role, clientType = clientType))))
+  }
+
+  @PostMapping("base-clients")
+  @ResponseStatus(HttpStatus.OK)
+  @PreAuthorize("hasRole('ROLE_OAUTH_ADMIN')")
+  fun addClient(
+    @Valid @RequestBody
+    clientDetails: ClientRegistrationRequest,
+  ): ResponseEntity<Any> {
+    val registrationResponse = saveClientService.addClient(clientDetails)
+    val telemetryMap = mapOf("username" to authenticationFacade.currentUsername!!, "clientId" to clientDetails.clientId!!)
+    telemetryClient.trackEvent("AuthorizationServerDetailsAdd", telemetryMap)
+    return ResponseEntity.ok(registrationResponse)
   }
 
   @GetMapping("clients/duplicates/{clientId}")
