+++ src/main/kotlin/uk/gov/justice/digital/hmpps/authorizationserver/resource/converter/ClientCredentialsViewResponseConverter.kt
@@ -2,15 +2,15 @@ package uk.gov.justice.digital.hmpps.authorizationserver.resource.converter
 
 import org.springframework.core.convert.converter.Converter
 import org.springframework.stereotype.Component
-import uk.gov.justice.digital.hmpps.authorizationserver.resource.ClientCredentialsViewResponse
+import uk.gov.justice.digital.hmpps.authorizationserver.resource.ClientViewResponse
 import uk.gov.justice.digital.hmpps.authorizationserver.service.ClientComposite
 
 @Component
-class ClientCredentialsViewResponseConverter : Converter<ClientComposite, ClientCredentialsViewResponse> {
+class ClientCredentialsViewResponseConverter : Converter<ClientComposite, ClientViewResponse> {
 
-  override fun convert(source: ClientComposite): ClientCredentialsViewResponse? {
+  override fun convert(source: ClientComposite): ClientViewResponse? {
     with(source) {
-      return ClientCredentialsViewResponse(
+      return ClientViewResponse(
         latestClient.clientId,
         latestClient.scopes,
         authorizationConsent?.authorities,
